move unescapeMuxVars to server
